refactor: refactor grpc clients list connection
