+++ src/grpc-clients/src/mqtt/admin/inner.rs
@@ -15,13 +15,11 @@
 use common_base::error::common::CommonError;
 use mobc::Connection;
 use prost::Message;
-use protocol::broker_mqtt::broker_mqtt_admin::{
-    ClusterStatusReply, ClusterStatusRequest, CreateUserReply, CreateUserRequest, DeleteUserReply,
-    DeleteUserRequest, ListUserReply, ListUserRequest,
-};
-
+use protocol::broker_mqtt::broker_mqtt_admin::{ClusterStatusReply, ClusterStatusRequest, CreateUserReply, CreateUserRequest, DeleteUserReply, DeleteUserRequest, ListConnectionReply, ListConnectionRequest, ListUserReply, ListUserRequest};
+use crate::mqtt::connection::MqttBrokerConnectionServiceManager;
 use super::MqttBrokerAdminServiceManager;
 
+// ---- cluster -------
 pub(crate) async fn inner_cluster_status(
     mut client: Connection<MqttBrokerAdminServiceManager>,
     request: Vec<u8>,
@@ -35,6 +33,7 @@ pub(crate) async fn inner_cluster_status(
     }
 }
 
+// ----- user -------
 pub(crate) async fn inner_list_user(
     mut client: Connection<MqttBrokerAdminServiceManager>,
     request: Vec<u8>,
@@ -68,6 +67,21 @@ pub(crate) async fn inner_delete_user(
     match DeleteUserRequest::decode(request.as_ref()) {
         Ok(request) => match client.mqtt_broker_delete_user(request).await {
             Ok(result) => Ok(DeleteUserReply::encode_to_vec(&result.into_inner())),
+            Err(e) => Err(CommonError::GrpcServerStatus(e)),
+        },
+        Err(e) => Err(CommonError::CommonError(e.to_string())),
+    }
+}
+
+
+// ------- connection ------
+pub(crate) async fn inner_list_connection(
+    mut client: Connection<MqttBrokerAdminServiceManager>,
+    request: Vec<u8>,
+) -> Result<Vec<u8>, CommonError> {
+    match ListConnectionRequest::decode(request.as_ref()) {
+        Ok(request) => match client.mqtt_broker_list_connection(request).await {
+            Ok(result) => Ok(ListConnectionReply::encode_to_vec(&result.into_inner())),
             Err(e) => Err(CommonError::GrpcServerStatus(e)),
         },
         Err(e) => Err(CommonError::CommonError(e.to_string())),
